[EngSys] Move context.js under workflows/src - Should only be needed by workflows - If needed by shared actions, workflows can pass specific inputs
